+++ src/control_plane/configuration/mod.rs
@@ -1,6 +1,7 @@
 mod apply_artifact_lock;
 mod artifact_lock;
 mod artifact_lock_error;
+mod artifact_source;
 mod config_parse_error;
 mod parse_artifact_lock;
 mod parse_project_config;
@@ -9,8 +10,9 @@ mod raw_project_config;
 mod raw_service_config;
 
 pub(crate) use apply_artifact_lock::apply_artifact_lock;
-pub(crate) use artifact_lock::ArtifactLock;
+pub(crate) use artifact_lock::{ArtifactLock, ArtifactLockImage};
 pub(crate) use artifact_lock_error::ArtifactLockError;
+pub(crate) use artifact_source::artifact_source;
 pub(crate) use config_parse_error::ConfigParseError;
 pub(crate) use parse_artifact_lock::parse_artifact_lock;
 pub(crate) use parse_project_config::parse_project_config;